+++ fort-to-nite/Viewcontrollers/DetailsItemViewController.swift
@@ -50,6 +50,7 @@ class DetailsItemViewController: UIViewController {
     var itemDetails = ItemsDetails()
     var itemModel = JsonService.list
     var likeButtonState: Bool?
+    var gl: CAGradientLayer!
 
     // MARK: - Outlets
     
@@ -99,13 +100,32 @@ class DetailsItemViewController: UIViewController {
             likeButtonState = false
         }
     }
+    
+    func viewDidLayoutSubviews(from fromInterfaceOrientation: UIInterfaceOrientation) {
+        backgroundGradient()
+    }
+
 
     // MARK: - Visuals
 
     private func backgroundGradient() {
-        BackgroundColors().backgroundGradient(view: view)
+        let colorTop = UIColor(red: 55.0 / 255.0, green: 194.0 / 255.0, blue: 254.0 / 255.0, alpha: 1.0).cgColor
+        let colorBottom = UIColor(red: 115.0 / 255.0, green: 50.0 / 255.0, blue: 252.0 / 255.0, alpha: 1.0).cgColor
+        
+        gl = CAGradientLayer()
+        gl.frame = view.bounds
+        gl.colors = [colorTop, colorBottom]
+        gl.locations = [0.0, 1.0]
+        gl.name = "background_gradient"
+        
+        view.backgroundColor = UIColor.clear
+        view.layer.insertSublayer(gl, at: 0)
     }
 
+    override func viewDidLayoutSubviews() {
+        gl.frame = view.bounds
+    }
+    
     private func getGradientValueForBackgroundImage() {
         FormatLevels().formatUIBackgroundViewFromLevel(view: backgroundImageView, level: itemInfo.color)
         shadows.setShadow(image: itemImage)