+++ fort-to-nite/Viewcontrollers/DetailsWeaponViewController.swift
@@ -22,6 +22,7 @@ class DetailsWeaponViewController: UIViewController {
     var weaponInfo = Weapons()
     var weaponDetails = WeaponsDetails()
     var weaponModel = JsonService.list
+    var gl: CAGradientLayer!
     
     struct likeActions {
         let like: Bool
@@ -151,6 +152,11 @@ class DetailsWeaponViewController: UIViewController {
         getLikeStorage()
     }
     
+    func viewDidLayoutSubviews(from fromInterfaceOrientation: UIInterfaceOrientation) {
+        backgroundGradient()
+    }
+
+    
     // MARK: - Logic
 
     private func prepareSegment() {
@@ -176,7 +182,21 @@ class DetailsWeaponViewController: UIViewController {
     // MARK: - Visuals
 
     private func backgroundGradient() {
-        BackgroundColors().backgroundGradient(view: view)
+        let colorTop = UIColor(red: 55.0 / 255.0, green: 194.0 / 255.0, blue: 254.0 / 255.0, alpha: 1.0).cgColor
+        let colorBottom = UIColor(red: 115.0 / 255.0, green: 50.0 / 255.0, blue: 252.0 / 255.0, alpha: 1.0).cgColor
+        
+        gl = CAGradientLayer()
+        gl.frame = view.bounds
+        gl.colors = [colorTop, colorBottom]
+        gl.locations = [0.0, 1.0]
+        gl.name = "background_gradient"
+        
+        view.backgroundColor = UIColor.clear
+        view.layer.insertSublayer(gl, at: 0)
+    }
+    
+    override func viewDidLayoutSubviews() {
+        gl.frame = view.bounds
     }
     
     private func prepareVisuals() {