+++ fort-to-nite/Viewcontrollers/ItemsViewController.swift
@@ -45,6 +45,7 @@ class ItemsViewController: UIViewController, UICollectionViewDelegate, UICollect
 
     var cellParentId: Int = 0
     var index: Int?
+    var gl: CAGradientLayer!
 
     @IBOutlet weak var collectionView: UICollectionView!
 
@@ -59,6 +60,10 @@ class ItemsViewController: UIViewController, UICollectionViewDelegate, UICollect
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(true)
         reloadData()
+    }
+    
+    override func viewDidLayoutSubviews() {
+        gl.frame = view.bounds
     }
 
     // MARK: - Data
@@ -107,6 +112,16 @@ class ItemsViewController: UIViewController, UICollectionViewDelegate, UICollect
     }
 
     private func backgroundGradient() {
-        BackgroundColors().backgroundGradient(view: view)
+        let colorTop = UIColor(red: 55.0 / 255.0, green: 194.0 / 255.0, blue: 254.0 / 255.0, alpha: 1.0).cgColor
+        let colorBottom = UIColor(red: 115.0 / 255.0, green: 50.0 / 255.0, blue: 252.0 / 255.0, alpha: 1.0).cgColor
+        
+        gl = CAGradientLayer()
+        gl.frame = view.bounds
+        gl.colors = [colorTop, colorBottom]
+        gl.locations = [0.0, 1.0]
+        gl.name = "background_gradient"
+        
+        view.backgroundColor = UIColor.clear
+        view.layer.insertSublayer(gl, at: 0)
     }
 }