+++ fort-to-nite/Viewcontrollers/WeaponViewController.swift
@@ -45,7 +45,6 @@ class WeaponCollectionViewFooterCell : UICollectionViewCell {
     func configure() {
         let shadowsOptions = ShadowLayers()
         shadowsOptions.setShadow(label: jsonVersion)
-        jsonVersion.font = UIFont(name: "BurbankBigCondensed-Bold", size: 19)
     }
     
 }
@@ -59,6 +58,7 @@ class WeaponViewController: UIViewController, UICollectionViewDelegate, UICollec
     let levels = FormatLevels()
     let isFirstLaunch = UserDefaults.isFirstLaunch()
     let updateWelcome = UserDefaults.lastUpdate()
+    var gl: CAGradientLayer!
     var cellParentId: Int = 0
     var index: Int?
 
@@ -123,6 +123,10 @@ class WeaponViewController: UIViewController, UICollectionViewDelegate, UICollec
     func reloadData() {
         collectionView?.reloadData()
     }
+    
+    override func viewDidLayoutSubviews() {
+        gl.frame = view.bounds
+    }
 
     // MARK: - Collection View
 
@@ -148,8 +152,11 @@ class WeaponViewController: UIViewController, UICollectionViewDelegate, UICollec
     
     func collectionView(_ collectionView: UICollectionView, viewForSupplementaryElementOfKind kind: String, at indexPath: IndexPath) -> UICollectionReusableView {
         let footer = collectionView.dequeueReusableSupplementaryView(ofKind: kind, withReuseIdentifier: "footer", for: indexPath) as! WeaponCollectionViewFooterCell
-        
-        footer.jsonVersion.text = "Updated for: \(list.getJsonVersion())"
+        let jsonVersion = list.getJsonVersion()
+        footer.configure()
+        if (list.getJsonVersion() != "") {
+            footer.jsonVersion.text = "Data pulled from: V\(jsonVersion)"
+        }
         return footer
     }
 
@@ -175,6 +182,16 @@ class WeaponViewController: UIViewController, UICollectionViewDelegate, UICollec
     // MARK: - Visuals
 
     private func backgroundGradient() {
-        BackgroundColors().backgroundGradient(view: view)
+        let colorTop = UIColor(red: 55.0 / 255.0, green: 194.0 / 255.0, blue: 254.0 / 255.0, alpha: 1.0).cgColor
+        let colorBottom = UIColor(red: 115.0 / 255.0, green: 50.0 / 255.0, blue: 252.0 / 255.0, alpha: 1.0).cgColor
+        
+        gl = CAGradientLayer()
+        gl.frame = view.bounds
+        gl.colors = [colorTop, colorBottom]
+        gl.locations = [0.0, 1.0]
+        gl.name = "background_gradient"
+        
+        view.backgroundColor = UIColor.clear
+        view.layer.insertSublayer(gl, at: 0)
     }
 }